requirements.txt
Django == 2.1.7
dj-database-url == 0.5.0
djangorestframework  == 3.10.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
whitenoise==5.2.0